LambdaExpression Tasks 1, 2 are done
