+++ src/main/java/com/rakovets/course/javabasics/practice/lambdaexpressions/CitiesFilterUtil.java
@@ -1,29 +1,32 @@
 package com.rakovets.course.javabasics.practice.lambdaexpressions;
 
-import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
 
 public class CitiesFilterUtil {
-    public static void main(String[] args) {
-        ArrayList<String> list = new ArrayList<>(){{
-            add("Belgrade");
-            add("Minsk  ");
-            add("Moscow");
-            add("Kiev");
-            add("London");
-            add("New York");
-            add("Sydney");
-            add("St.Petersburg");
-            add("Dresden");
-            add("Hamburg");
-            add("Berlin");
-            add("Potsdam");
-            add("Essen");
-            add("Leipzig");
-            add("Stuttgart");
-            add("Konigsberg");
-            add("Minsk");
-        }};
-        list.stream()
-                .forEach(System.out::println);
-         }
+
+    public static List<String> getDistinctCitiesInList(List<String> list) {
+        return list.stream()
+                .distinct()
+                .collect(Collectors.toList());
+    }
+
+    public static List<String> getCitiesLessThan6LettersNames(List<String> list) {
+        return list.stream().
+                filter(city -> city.length() > 6).
+                collect(Collectors.toList());
+    }
+
+    public static List<String> getCitiesNamesStartedWithLetter(List<String> list, char letter) {
+        return list.stream().
+                distinct().
+                filter(city -> city.charAt(0) == letter).
+                collect(Collectors.toList());
+    }
+
+    public static long countDistinctCity(List<String> list, String city) {
+        return list.stream().
+                filter(cities -> cities.equals(city)).
+                count();
+    }
 }
